# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tropia Live API
  version: 1.0.0
  description: |
    REST API for the Tropia Live shopping platform.

    All authenticated endpoints require a `Authorization: Bearer <jwt>`
    header. Refresh tokens live in the HttpOnly `refresh_token` cookie
    set on `/api/auth/login` and rotated on `/api/auth/refresh`.

    **OWASP API Security Top 10 compliance:**
    - API1 (BOLA): ownership checks on every mutate endpoint.
    - API2 (Broken Auth): bcrypt cost 12, refresh rotation + reuse
      detection, sliding-window rate-limit (Redis Lua).
    - API3 (Mass Assignment): registration role is whitelisted to
      buyer/seller only; admin must be promoted DB-side.
    - API4 (Resource Consumption): pagination caps, 5MB upload cap,
      DeepSeek bot quota 30/hour/session.
    - API5 (BFLA): admin/seller middleware applied per route.
    - API6 (Sensitive Flow): orders 20/min/user, like 60/min/IP, chat
      30/min/user, login 5/min/IP.
    - API8 (Misconfig): CORS whitelist, security headers, GIN_MODE
      enforcement, dev-only fields gated.
    - API9 (Inventory): this spec.
    - API10 (Unsafe Consumption): HTTP client timeouts on outbound
      MoMo/ZaloPay/DeepSeek calls.
  contact:
    name: Tropia Backend Team
servers:
  - url: https://api.tropia.vn
    description: Production
  - url: http://localhost:3000
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        error:
          type: string
          example: email format invalid
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [buyer, seller, admin]
        avatar_url:
          type: string
          format: uri
          nullable: true
        phone:
          type: string
          nullable: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Must contain uppercase + digit.
        name:
          type: string
        phone:
          type: string
        shop_name:
          type: string
        role:
          type: string
          enum: [buyer, seller]
          description: |
            Self-service can only request buyer or seller. Any other
            value (including `admin`) is silently coerced to `buyer`.
            This is the OWASP API3 (Mass Assignment) defense — admins
            must be promoted via DB-side migration, never via API.
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Create a new buyer or seller account
      description: |
        Rate-limited to 3 requests / 5 min / IP. OTP is emailed
        asynchronously; account is locked until `/verify-otp` succeeds.
        Role field is whitelisted — `admin` requests are coerced to
        `buyer`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created. OTP sent via email.
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Email already registered
        '429':
          description: Too many requests (3/5min)
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Exchange credentials for JWT + refresh cookie
      description: Rate-limited to 5/min/IP. Locks account after 5 failed attempts for 15 min.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Access token + user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          description: Invalid credentials
        '423':
          description: Account locked (too many failed attempts)
        '429':
          description: Rate limited
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token, mint new access token
      description: |
        Reads the refresh token from the `refresh_token` cookie. If the
        same refresh token is presented twice (reuse), the entire token
        family is revoked — defense against stolen refresh tokens.
      security: []
      responses:
        '200':
          description: New access + refresh
        '401':
          description: Invalid/expired refresh
...
  /api/upload/avatar:
    post:
      tags: [Upload]
      summary: Upload user avatar (5MB max, jpg/png/webp)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: R2 URL
  /api/upload/product:
    post:
      tags: [Upload]
      summary: Upload product images (seller only, owner-only)
      description: |
        Verifies the authenticated seller owns `product_id` before
        accepting the upload — OWASP API1 (BOLA) defense.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [product_id, images]
              properties:
                product_id:
                  type: string
                  format: uuid
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: URLs
        '404':
          description: Product not yours
...